requirements.txt
fabric-cicd>=0.1.0
azure-identity>=1.15.0
pyyaml>=6.0
